2026.08.24 매장 관리 API 추가 / 관리자 등록 시 매장 설정 및 카테고리 삭제 시 해당 카테고리 사용 매장 조회 추가
- 등록
- 수정
- 삭제
- 활성화 / 비활성화
- 리스트 (페이지)
- 리스트 (전체)
- 정보

+++ src/main/java/com/project/dine/reserve/service/store/StoreCategoryService.java
@@ -5,6 +5,7 @@ import com.project.dine.reserve.domain.store.DineReserveStoreCategory;
 import com.project.dine.reserve.dto.constant.error.StoreErrorCode;
 import com.project.dine.reserve.dto.store.category.*;
 import com.project.dine.reserve.repository.store.DineReserveStoreCategoryRepository;
+import com.project.dine.reserve.repository.store.DineReserveStoreInfoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -20,6 +21,7 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class StoreCategoryService {
     private final DineReserveStoreCategoryRepository dineReserveStoreCategoryRepository;
+    private final DineReserveStoreInfoRepository dineReserveStoreInfoRepository;
 
     @Transactional
     public void storeCategoryRegist(StoreCategoryRegist storeCategoryRegist) {
@@ -44,7 +46,8 @@ public class StoreCategoryService {
         DineReserveStoreCategory dineReserveStoreCategory = dineReserveStoreCategoryRepository.findByCategoryUUID(categoryUUID)
                 .orElseThrow(() -> new DineReserveException(StoreErrorCode.NO_STORE_CATEGORY));
 
-        // todo 해당 카테고리에 존재하는 매장이 있는지 check
+        int count = dineReserveStoreInfoRepository.countByCategoryUUID(dineReserveStoreCategory.getCategoryUUID());
+        if (count > 0) throw new DineReserveException(StoreErrorCode.USE_CATEGORY);
 
         dineReserveStoreCategoryRepository.delete(dineReserveStoreCategory);
     }
